Made player directions work, I think
However, the camera is broken and I don't entirely know why.

+++ battle of the aces/src/main/main.java
@@ -74,11 +74,11 @@ public class main extends Canvas implements Runnable{
 		g.fillRect(0, 0, width, height);
 		
 		if (renderer.player != null) {
-			g.translate(-renderer.player.getX()+renderer.player.getWidth(), -renderer.player.getY()+renderer.player.getHeight());
+			g.translate(-(int)renderer.player.getX()+renderer.player.getWidth(), (int)-renderer.player.getY()+renderer.player.getHeight());
 			
 			renderer.render(g);
 			
-			g.translate(renderer.player.getX()+renderer.player.getWidth(), renderer.player.getY()+renderer.player.getHeight());
+			g.translate((int)renderer.player.getX()+renderer.player.getWidth(), (int)renderer.player.getY()+renderer.player.getHeight());
 		}else {
 			renderer.render(g);
 		}
